+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.0.4] - In progress
+### Added
+- Helper for counting group sizes in pandas
+
 ### Updated
 - xgb_utils to be compatible with `xgboost > 1.0`.